requirements.txt
langchain
langchain-community
langgraph
python-dotenv
pydantic
rich
portkey_ai
langchain_openai
duckdb